# Use `PagedAllocator` for `StringName` to accelerate and localize allocations.