Requires https://github.com/godotengine/godot/pull/105760 (out of convenience)

Alternative to https://github.com/godotengine/godot/pull/104372 and https://github.com/godotengine/godot/pull/97016.
This PR is a much less extreme change, so I'd prefer if it was merged first and the other PRs tested against master with a paged allocator.

I propose to change `memnew` calls in `StringName` with a dedicated `PagedAllocator`. This should accelerate allocation and deallocation and reduce used RAM.

https://github.com/godotengine/godot/pull/97016 is not superseded; but it would have to prove it's substantially better than the paged allocator, not just better than the default dynamic allocator.

## Benchmark

I benchmarked a ~6% acceleration for `Node` creation and deletion:
```c++
	Node *parent = memnew(Node);

	{
		auto t0 = std::chrono::high_resolution_clock::now();
		for (int i = 0; i < 2000000; i ++) {
			Node *child = memnew(Node);
			parent->add_child(child);
		}
		auto t1 = std::chrono::high_resolution_clock::now();
		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
	}
	{
		Array children = parent->get_children();
		auto t0 = std::chrono::high_resolution_clock::now();
		for (int i = 0; i < 2000000; i ++) {
			parent->remove_child((Node *)(Object *)children.get(i));
		}
		auto t1 = std::chrono::high_resolution_clock::now();
		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
	}
```

This yields:
```
2364ms
708ms
```
(master)

```
2240ms
682ms
```
(this PR)

I also tested a ~13mb decrease of RAM use:
- `master` uses `1540705200b`
- this PR uses `1527547728b`